requirements.txt
# =============================================================
# requirements.txt — VisionDX Production Dependencies
# =============================================================

# ─── Web Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
python-multipart==0.0.9
httpx==0.27.0

# ─── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# ─── Auth / Security ──────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5

# ─── PDF / OCR ────────────────────────────────────────────────
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.13
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# ─── ML / Data Science ────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0
imbalanced-learn==0.12.2
pyarrow==16.1.0

# ─── NLP ──────────────────────────────────────────────────────
rapidfuzz==3.9.0

# ─── Voice (optional, for /chat/voice) ─────────────────────────
# pip install SpeechRecognition  # uses Google STT; supports all languages

# ─── Config / Utilities ───────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
loguru==0.7.2
tenacity==8.3.0
